+++ t/09-prompt-alignment.t
@@ -9,7 +9,7 @@ use Test::More;
 
 our $VERSION = '0.001';
 
-const my $EXPECTED_TESTS => 42;
+const my $EXPECTED_TESTS => 45;
 
 plan tests => $EXPECTED_TESTS;
 
@@ -91,6 +91,9 @@ for my $required_term (
     SubscriptionStore
     PreferenceStore
     NotificationDispatcher
+    DocumentBuilder
+    SearchIndexer
+    Searcher
     post_bodies
     thread_counters
     thread_counter_shards