+++ t/lib/GPForum/Test/SearchRow.pm
@@ -1,0 +1,28 @@
+package GPForum::Test::SearchRow;
+
+use strict;
+use warnings;
+
+use Mojo::Base -base;
+
+our $VERSION = '0.001';
+
+has category     => undef;
+has current_body => undef;
+has data         => sub { return {}; };
+has thread       => undef;
+
+sub get_column {
+    my ( $self, $name ) = @_;
+
+    return $self->data->{$name};
+}
+
+sub has_column {
+    my ( $self, $name ) = @_;
+
+    return exists $self->data->{$name};
+}
+
+1;
+